src/lib.c: avoid DAC_OVERRIDE denial and use root:snap_daemon
tcpdump -w <file> causes tcpdump to do a chown(path, "tcpdump", "tcpdump")
which causes a DAC_OVERRIDE denial. So we instead set the capture directory
to 0770 with root:snap_daemon and then set these files as root:snap_daemon,
which allows the initial write but also -w to overwrite the same file. When
using -Z root, we see chown on root:root, so honor that.

For now, this works ok with 0770 root:snap_daemon capture directory, but
the files are 0644. We might want to set the umask to 0002 instead so the
files of 0664.