use -Z snap_daemon via wrapper instead of modifying setgid/setuid
Changing setgid() and setuid() via LD_PRELOAD worked for those
functions, but tcpdump by default does a getent on the tcpdump
user/group and this fails on systems without that user/group (like UC),
so instead use a wrapper to detect '-Z root' and if not specified, add
'-Z snap_daemon'. This is imperfect since if the user does '-Z foo',
then ps will show '-Z foo ... -Z snap_daemon', however it is functional
and means we don't have to account for -Zfoo or -Z foo to replace the
value. Perhaps fix this in the future, but for now it is fine.